Delete mathematics.cpp/h
