+++ UWP/CommonUWP/CommonUWP.vcxproj.filters
@@ -492,9 +492,6 @@
     <ClCompile Include="..\..\ext\at3_standalone\get_bits.cpp">
       <Filter>ext\at3_standalone</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\ext\at3_standalone\mathematics.cpp">
-      <Filter>ext\at3_standalone</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\ext\at3_standalone\mem.cpp">
       <Filter>ext\at3_standalone</Filter>
     </ClCompile>
@@ -947,9 +944,6 @@
       <Filter>ext\at3_standalone</Filter>
     </ClInclude>
     <ClInclude Include="..\..\ext\at3_standalone\intreadwrite.h">
-      <Filter>ext\at3_standalone</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\ext\at3_standalone\mathematics.h">
       <Filter>ext\at3_standalone</Filter>
     </ClInclude>
     <ClInclude Include="..\..\ext\at3_standalone\mem.h">
