+++ android/jni/Android.mk
@@ -122,7 +122,6 @@ AT3_STANDALONE_FILES := \
 	${SRC}/ext/at3_standalone/compat.cpp \
 	${SRC}/ext/at3_standalone/fft.cpp \
 	${SRC}/ext/at3_standalone/float_dsp.cpp \
-	${SRC}/ext/at3_standalone/mathematics.cpp \
 	${SRC}/ext/at3_standalone/mem.cpp \
 	${SRC}/ext/at3_standalone/sinewin.cpp
