+++ ext/at3_standalone/CMakeLists.txt
@@ -15,7 +15,6 @@ set(ALL_SOURCE_FILES
 	${SRC_DIR}/compat.cpp
 	${SRC_DIR}/fft.cpp
 	${SRC_DIR}/float_dsp.cpp
-	${SRC_DIR}/mathematics.cpp
 	${SRC_DIR}/mem.cpp
 	${SRC_DIR}/sinewin.cpp
 	)
