+++ ext/at3_standalone/atrac3plusdsp.cpp
@@ -41,6 +41,7 @@
 
 #endif
 
+#define _USE_MATH_DEFINES
 #include <math.h>
 #include <string.h>
