+++ ext/at3_standalone/float_dsp.h
@@ -27,23 +27,6 @@ void vector_fmul(float *dst, const float *src0, const float *src1, int len);
 void vector_fmul_scalar(float *dst, const float *src, float mul, int len);
 
 /**
-    * Calculate the entry wise product of two vectors of floats, add a third vector of
-    * floats and store the result in a vector of floats.
-    *
-    * @param dst  output vector
-    *             constraints: 32-byte aligned
-    * @param src0 first input vector
-    *             constraints: 32-byte aligned
-    * @param src1 second input vector
-    *             constraints: 32-byte aligned
-    * @param src2 third input vector
-    *             constraints: 32-byte aligned
-    * @param len  number of elements in the input
-    *             constraints: multiple of 16
-    */
-void vector_fmul_add(float *dst, const float *src0, const float *src1, const float *src2, int len);
-
-/**
     * Calculate the entry wise product of two vectors of floats, and store the result
     * in a vector of floats. The second vector of floats is iterated over
     * in reverse order.
