+++ libretro/Makefile.common
@@ -272,7 +272,6 @@ SOURCES_CXX += \
 	${EXTDIR}/at3_standalone/compat.cpp \
 	${EXTDIR}/at3_standalone/fft.cpp \
 	${EXTDIR}/at3_standalone/float_dsp.cpp \
-	${EXTDIR}/at3_standalone/mathematics.cpp \
 	${EXTDIR}/at3_standalone/mem.cpp \
 	${EXTDIR}/at3_standalone/sinewin.cpp
